better take on constructing cmdline var
So the problem looks like this:

- We want to populate a var with the command line used to launch Quake.

- We are trying to reconstruct that command line from the argument vector.

- Proper reconstruction depends on the arg-quoting conventions in the environment used to launch Quake.

In the spirit of minimal churn to the existing code, what I've settled on is:

- COM_InitArgv will use a bash-like quoting convention for command-line reconstruction. Pretty simple stuff: quote arguments that contain whitespace, backslash-quote any actual quote chars, and backslash-quote backslash chars as needed.

- If the caller doesn't want that convention (which, in Windows, we don't) then the caller can provide the actual command line if that's available, to be used instead of reconstructing from the arg vector. On Windows, the command line is in fact available, so provide it.

I've been using this with the Windows build... works fine, populates the cmdline var as expected.

The Linux build compiles and starts, but I don't have a graphical Linux environment yet that can run reQuiem (fails with XF86DGANoDirectVideoMode) so you'll probably want to sanity-test this on Linux still.