fix listen-server crash
This is a fix for issue #8. The comments on that issue describe the problem.
This change fixes that issue and is also just generally correct in that we
shouldn't be drawing scoreboards if we aren't even signed into the game yet.

However it would be cool if a Quake engine coding veteran could verify that
this particular change is the best (least fragile) way of addressing the issue.
I've chosen the "bail out at the last minute" approach, and it might be that
a higher-level fix to the control flow would be better.