don't generate bad svc_spawnstatic commands at changelevel
Resolves issue #16.

PF_makestatic has some additional code for supporting the "create" command.
However, this can incorrectly run during a changelevel (because the
client spawned flag is true), pushing extraneous svc_spawnstatic commands to
the client.

The client will parse the commands incorrectly because it hasn't yet
received the svc_serverinfo command that sets the protocol version. This
results in some "command parsing roulette" which may or may not cause the
level load to fail.

Adding this global flag check ensures that the extra code only runs, as
intended, during the create command. I left the other checks in place just to
be safe... the create command is a test/debug command in any case so a few
extra checks are not a performance concern.

The test cases of the zendar and masque map changes work now. The create
command still works too.