change makestatic check from "allow_precache" to "sv.state == ss_active"
I like this better. Checking sv.state most accurately excludes the time during
which this would be a bad idea (and hints at why).

If we were to be the most precise & complete about the bit of functionality in
this function, we should probably send the svc_spawnstatic message to all
clients. Leave it up to Host_Create_f to decide if the "create" command
should be limited to the single-local-client case or not. But, let's do
minimal changes for now and just fix the bug.