re-enabling crosshair display during demo playback
jdhack preferred no crosshair being shown when playing back a demo and hard-coded that. I am using my new power to revert that to the original standard. :)

I think this is better, but if someone wants make a cvar for this (cl_crosshair_showduringdemo or something) feel welcome! Maybe I will do that for practice some day.