after creating a compressed demo, add the dzip to com_searchpaths
Resolves issue #30.

com_searchpaths has existing .dz files added to it at startup time. If we
create a new .dz file we should add it too.

Note that we'll only add .dz files (not .zip files) to the searchpaths, to
mimic the startup-time behavior which also only looks for .dz files.

Tested by doing demo compress with the dzip app and/or library installed.