always-mouselook cvar
A possible solution for issue #55.

This adds the cvar m_look, which (as in Engoo) changes the global default for
mouselook. m_look 0 behaves like original Quake. m_look 1 flips things around;
in that situation mouselook will be enabled UNLESS you are pressing a +mlook key.

m_look is also exposed in the menu as "Always Mouselook".

m_look is not enabled by default. I waffled on this, but the downside of
enabling it by default is that if you have the habit of using an autoexec
file with "+mlook" in it, then you would get the opposite of what you expect.
Since it's exposed in the menu, maybe it's fine not to enable it by default?

Alternatively we could make m_look behave a little differently than in Engoo:
instead of flipping the sense of the +mlook key, it could just override it.
E.g. mouselook would be enabled if EITHER m_look is 1 or +mlook is pressed.
That would be an easy 1-line change to the new RecalcMouselook function.
I'm starting to lean that way.